API Refactor: carto.dataview
