+++ src/lib/dataview/category/CategoryDataViewImpl.ts
@@ -3,7 +3,7 @@ import { DataViewMode, DataViewCalculation, DataViewData } from '../mode/DataVie
 import { DataViewImpl } from '../DataViewImpl';
 import { CartoDataViewError, dataViewErrorTypes } from '../DataViewError';
 
-export class CategoryImpl<T extends DataViewMode> extends DataViewImpl<T> {
+export class CategoryDataViewImpl<T extends DataViewMode> extends DataViewImpl<T> {
   public operationColumn: string;
   public limit?: number;
