+++ src/lib/dataview/formula/FormulaDataViewImpl.ts
@@ -2,7 +2,7 @@ import { AggregationType } from '../../data/operations/aggregation/aggregation';
 import { DataViewData, DataViewMode } from '../mode/DataViewMode';
 import { DataViewImpl } from '../DataViewImpl';
 
-export class FormulaImpl<T extends DataViewMode> extends DataViewImpl<T> {
+export class FormulaDataViewImpl<T extends DataViewMode> extends DataViewImpl<T> {
   public async getData(): Promise<Partial<DataViewData>> {
     let formulaResponse;
